# src/openapi.yaml
openapi: 3.1.0
info:
  title: Volvix POS API
  description: |
    API serverless de Volvix POS conectada a Supabase. Multi-tenant, multi-rol (cajero, admin, owner, superadmin).
    Autenticación basada en JWT (Bearer). Todas las rutas bajo `/api/*` salvo `/api/login`, `/api/logout`,
    `/api/health`, `/api/health/deep`, `/api/errors/log`, y los endpoints de password-reset requieren `Authorization: Bearer <jwt>`.
  version: 7.3.0-r14
  contact:
    name: Volvix POS
    url: https://volvix-pos.vercel.app
  license:
    name: Proprietary
servers:
  - url: https://volvix-pos.vercel.app
    description: Producción
  - url: http://localhost:3000
    description: Desarrollo local
security:
  - BearerAuth: []
...
paths:
...
  /api/status:
    get:
      tags: [Health]
      summary: Estado de cableado (status.json)
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: object, additionalProperties: true } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { description: status.json no encontrado, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        '500': { $ref: '#/components/responses/ServerError' }
...
  /api/owner/dashboard:
    get:
      tags: [Owner]
      summary: Dashboard agregado (admin/owner/superadmin)
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/OwnerDashboard' } } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
...
    Unauthorized:
      description: No autenticado o JWT inválido
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Forbidden:
      description: Sin permisos para esta operación
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
    ServerError:
      description: Error interno
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
...
    Tenant:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        owner_user_id: { type: string, format: uuid, nullable: true }
        plan: { type: string, enum: [trial, free, pro, enterprise] }
        is_active: { type: boolean }
        created_at: { type: string, format: date-time }
...
    OwnerDashboard:
      type: object
      properties:
        ok: { type: boolean }
        metrics:
          type: object
          properties:
            total_users: { type: integer }
            active_users: { type: integer }
            total_tenants: { type: integer }
            active_tenants: { type: integer }
            total_sales: { type: integer }
            total_revenue: { type: number }
            total_products: { type: integer }
            low_stock_count: { type: integer }
            total_customers: { type: integer }
            active_customers: { type: integer }
            mrr: { type: number }
            arr: { type: number }
        sales_by_day: { type: array, items: { type: object, additionalProperties: true } }
        top_tenants: { type: array, items: { $ref: '#/components/schemas/Tenant' } }
...